+++ next-app/src/app/about/layout.tsx
@@ -8,7 +8,7 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <div>
+    <div className="container mx-auto px-4 py-8">
       <AboutPageComponent />
       {children}
       <LastUpdated date="11-11-2024" />
